+++ Valheim.SpawnThat/Spawners/SpawnerCreatureSpawner/CreatureSpawnerConfigManager.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using Valheim.SpawnThat.Configuration;
 using Valheim.SpawnThat.Core;
-using Valheim.SpawnThat.Reset;
 using Valheim.SpawnThat.Spawners.Caches;
 using Valheim.SpawnThat.Spawners.SpawnerCreatureSpawner.Types;
+using Valheim.SpawnThat.Startup;
 
 namespace Valheim.SpawnThat.Spawners.SpawnerCreatureSpawner
 {
