+++ Valheim.SpawnThat/Spawners/SpawnerSpawnSystem/Managers/SpawnPositionManager.cs
@@ -5,9 +5,9 @@ using System.Threading.Tasks;
 using UnityEngine;
 using Valheim.SpawnThat.Configuration.ConfigTypes;
 using Valheim.SpawnThat.Core;
-using Valheim.SpawnThat.Reset;
 using Valheim.SpawnThat.Spawners.Caches;
 using Valheim.SpawnThat.Spawners.SpawnerSpawnSystem.Position;
+using Valheim.SpawnThat.Startup;
 
 namespace Valheim.SpawnThat.Spawners.SpawnerSpawnSystem.Managers
 {
