+++ Valheim.SpawnThat/Spawners/SpawnerSpawnSystem/Managers/SpawnSystemConfigManager.cs
@@ -5,8 +5,8 @@ using Valheim.SpawnThat.Configuration;
 using Valheim.SpawnThat.Configuration.ConfigTypes;
 using Valheim.SpawnThat.Core;
 using Valheim.SpawnThat.Debugging;
-using Valheim.SpawnThat.Reset;
 using Valheim.SpawnThat.Spawners.Caches;
+using Valheim.SpawnThat.Startup;
 using Valheim.SpawnThat.Utilities;
 using Valheim.SpawnThat.Utilities.Extensions;
